Fixed poll colors hiding option text in light mode.
When a device's theme is set to light mode, if a poll option's text
is longer than the corresponding option's vote bar, the end of the
option text is not visible.

This changes the colour scheme used to more closely match how Twitter
iself renders the poll options, allowing the options to be more
visible in both light and dark mode.

Fixes #135.