Now hiding the donate button when building for the Play Store.
This is required because Google are evil, and will reject the app
if you link to a place where payments could be made, and those
payments are not made through Google's payment systems.

For apps to receive money, they must use Google's billing library,
so Fritter may be forced to accept donations through IAP in the
Play Store version in the future (including giving Google a 30% cut).