Fixed null pointer when loading a group feed.
This was introduced in 1297b1f85850ec698bfebe9acc4095746211f0f2, and
causes a null pointer to occur when opening a group feed. It happens
because when loading a group feed, a partial select from
`TABLE_SUBSCRIPTIONS` is performed, missing out the new `verified`
column.

The `verified` column is now included in the select, as a wildcard is
being used for now.