Fixed the "last hello build" value being set even on errors
The "hello" functionality sets a preference with the last version a
hello was sent from, to (naively) prevent spamming from one person
on the same version.

It was being set even when the hello failed, meaning a user couldn't say
hello until the next update. This fixes that, and now only sets the
preference if a successful hello occurs.