Replaced naive built-in logger with a real logger with levels.
This is more of an internal change that replaces the built-in Dart
log method with calls to a slightly more rich logger package, so
that we have support for logging things at different levels.

This is a stepping stone to having an in-app error log, and being able
to submit warnings and errors to the developers more easily.