Trying to fix subscriptions not being visible
This duplicates migration 14, from scratch as it looks like it hasn't
applied for some people, and I'm not sure why.

An error is being displayed on the subscriptions screen that implies the
new 'verified' column does not exist, but if the migrations failed to
run, it should essentially brick the app with an error (and it doesn't,
it seems). So this change removes the column and re-adds it, which might
solve the issue?