Fixed not being able to import recently exported data
This was caused by the addition of the 'verified' column in v2.10.0. The
files exported in that version used a boolean to hold the value, but
SQLite only supports inserting an integer, so it failed to cast.

This fixes the root cause, by now exporting an integer, and also adds in
temporary logic to handle importing the broken boolean versions.

Fixes #155.