Fixed a logging error that occurred on startup for some users.
Some users would get an error about "Controller is already firing an
event". This was caused by Fritter attempting to report an error from
inside the error listener, and the error listening library itself trying
to log, causing the listener to error as it doesn't seem to allow
concurrent access.