Fixed loading profiles that do not exist for various reasons
Some profiles are not able to be loaded for a multitude of reasons (e.g.
suspension, deactivation, deletion, etc.), and Fritter currently just
fails hard on these profiles.

This adds a customised error screen that displays a friendly message for
any supported error codes, and the actual error message coming back from
Twitter. It also logs unsupported error codes and messages so support
can be added for any that are missing (if the user opts in to Sentry
reporting).

Fixes #174.