Added specific error widget for socket errors
A lot of error reports come in for the app not being able to talk to
Twitter, but this isn't something that I can fix. I'm assuming the error
displayed isn't clear enough, so this change adds a dedicated error
widget whenever Fritter can't talk to Twitter, with a short summary that
the app "can't contact Twitter".