Moved save and share buttons into a context menu to save space
This moves the Save/Unsave and Share buttons into a context menu,
displayed at the top right of tweets. When tapped, the menu button opens
up a sheet modal at the bottom of the screen, just like in the video
player.

This will help with screen space on narrower devices, and should also
fix an annoying bug to do with unbound scrollbars.