Fixed truncated filenames when saving media (#255)
This was caused by some code in the file_picker_writable library we're
using to display the file picker. The code truncates the file name if
it's longer than 30 characters and I'm not sure why.

I fixed this by swapping to a different file picker library called
flutter_file_dialog.